Rework resource class to support using modules for retrieving resources by way of a `resource_retrieval_strategy`
Removes BasicResource class and replaces ActiveRelationResource with a module
