+++ lib/jsonapi-resources.rb
@@ -1,9 +1,8 @@
 require 'jsonapi/resources/railtie'
 require 'jsonapi/naive_cache'
 require 'jsonapi/compiled_json'
-require 'jsonapi/basic_resource'
-require 'jsonapi/active_relation_resource'
 require 'jsonapi/resource'
+require 'jsonapi/active_relation_retrieval_v10'
 require 'jsonapi/cached_response_fragment'
 require 'jsonapi/response_document'
 require 'jsonapi/acts_as_resource_controller'
